Update pylint configuration
