+++ tests/test_fixtures.py
@@ -1,6 +1,3 @@
-# pylint: disable=unused-import
-
-
 try:
     from _pytest.main import EXIT_OK  # type: ignore
 except ImportError:
